requirements.txt
fastapi
uvicorn
python-dotenv
python-jose 
msal
requests